+++ Cargo.toml
@@ -52,7 +52,6 @@ assert_matches = "1.5.0"
 async-trait = "0.1.68"
 base64 = "0.21.2"
 blockifier = { git = "https://github.com/starkware-libs/blockifier", tag = "v0.4.0-rc9.2" }
-cairo-lang-casm = "2.4.0"
 cairo-lang-compiler = "2.4.0"
 cairo-lang-debug = "2.4.0"
 cairo-lang-defs = "2.4.0"
@@ -79,12 +78,10 @@ camino = { version = "1.1.2", features = [ "serde1" ] }
 chrono = { version = "0.4.24", features = [ "serde" ] }
 clap = { version = "4.2", features = [ "derive" ] }
 clap_complete = "4.3"
-colored = "2"
 console = "0.15.7"
 convert_case = "0.6.0"
 crypto-bigint = { version = "0.5.3", features = [ "serde" ] }
 derive_more = "0.99.17"
-env_logger = "0.10.0"
 flate2 = "1.0.24"
 futures = "0.3.28"
 hex = "0.4.3"
@@ -92,7 +89,6 @@ indoc = "1.0.7"
 itertools = "0.10.3"
 lazy_static = "1.4.0"
 metrics-process = "1.0.9"
-num-bigint = "0.4"
 once_cell = "1.0"
 parking_lot = "0.12.1"
 pretty_assertions = "1.2.1"
